[memory][refactor] Split memory contexts and its access levels
