+++ sc-memory/sc-core/sc-store/sc_iterator5.c
@@ -8,6 +8,7 @@
 
 #include "../sc_memory_context_manager.h"
 #include "../sc_memory_context_private.h"
+#include "../sc_memory_context_access_levels.h"
 
 #include "sc-base/sc_allocator.h"
 
@@ -615,7 +616,8 @@ sc_bool sc_iterator5_next_ext(sc_iterator5 * it, sc_result * result)
     return status;
   }
 
-  if (_sc_memory_context_check_access_levels(sc_memory_get_context_manager(), it->ctx, SC_CONTEXT_ACCESS_LEVEL_READ)
+  if (_sc_memory_context_check_global_access_levels(
+          sc_memory_get_context_manager(), it->ctx, SC_CONTEXT_ACCESS_LEVEL_READ)
       == SC_FALSE)
   {
     *result = SC_RESULT_ERROR_SC_MEMORY_CONTEXT_HAS_NO_READ_ACCESS_LEVELS;
